Add a --host-state parameter
Add the values for the host states to the alert_colors, and work this
value in to the message.

Icinga sends notifications out about services and hosts, and this
command is used for both, so use --host-state as an indicator of
whether this is about a host in general, or a specific service.